Move CheckHasNoAssumesVisitor out of scope and fix it using ISinglePassCompiler
